added refactor code and error handler 404
