+++ src/main/java/com/jsr_dev/medical_api/controller/PhysicianController.java
@@ -1,6 +1,6 @@
 package com.jsr_dev.medical_api.controller;
 
-import com.jsr_dev.medical_api.physician.*;
+import com.jsr_dev.medical_api.domain.physician.*;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
