+++ src/main/java/com/jsr_dev/medical_api/domain/address/AddressResponse.java
@@ -1,4 +1,4 @@
-package com.jsr_dev.medical_api.address;
+package com.jsr_dev.medical_api.domain.address;
 
 public record AddressResponse(
         String street,
