+++ src/main/java/com/jsr_dev/medical_api/domain/patient/PatientMapper.java
@@ -1,31 +1,29 @@
-package com.jsr_dev.medical_api.physician;
+package com.jsr_dev.medical_api.domain.patient;
 
-import com.jsr_dev.medical_api.address.AddressMapper;
+import com.jsr_dev.medical_api.domain.address.AddressMapper;
 
-public class PhysicianMapper {
-    public static Physician mapToPhysician(AddPhysicianRequest p) {
-        return new Physician(
+public class PatientMapper {
+    public static Patient mapToPatient(AddPatientRequest p) {
+        return new Patient(
                 null,
                 p.name(),
                 p.avatar(),
                 p.email(),
+                p.identityDocument(),
                 p.phoneNumber(),
-                p.document(),
-                p.specialty(),
                 AddressMapper.mapToAddress(p.addAddressRequest()),
                 true
         );
     }
 
-    public static PhysicianResponse mapToPhysicianResponse(Physician p) {
-        return new PhysicianResponse(
+    public static PatientResponse mapToPatientResponse(Patient p) {
+        return new PatientResponse(
                 p.getId(),
                 p.getName(),
                 p.getAvatar(),
                 p.getEmail(),
+                p.getIdentityDocument(),
                 p.getPhoneNumber(),
-                p.getSpecialty(),
-                p.getDocument(),
                 AddressMapper.mapToAddressResponse(p.getAddress())
         );
     }
