+++ src/main/java/com/jsr_dev/medical_api/domain/patient/PatientRepository.java
@@ -1,4 +1,4 @@
-package com.jsr_dev.medical_api.patient;
+package com.jsr_dev.medical_api.domain.patient;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
