+++ src/main/java/com/jsr_dev/medical_api/domain/patient/UpdatePatientRequest.java
@@ -1,7 +1,7 @@
-package com.jsr_dev.medical_api.patient;
+package com.jsr_dev.medical_api.domain.patient;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
-import com.jsr_dev.medical_api.address.AddAddressRequest;
+import com.jsr_dev.medical_api.domain.address.AddAddressRequest;
 import jakarta.validation.constraints.NotNull;
 
 public record UpdatePatientRequest(
