+++ src/main/java/com/jsr_dev/medical_api/domain/physician/PhysicianResponse.java
@@ -1,6 +1,6 @@
-package com.jsr_dev.medical_api.physician;
+package com.jsr_dev.medical_api.domain.physician;
 
-import com.jsr_dev.medical_api.address.AddressResponse;
+import com.jsr_dev.medical_api.domain.address.AddressResponse;
 
 public record PhysicianResponse(
         Long id,
