+++ src/main/java/com/jsr_dev/medical_api/domain/physician/PhysicianResponseModelAssembler.java
@@ -1,4 +1,4 @@
-package com.jsr_dev.medical_api.physician;
+package com.jsr_dev.medical_api.domain.physician;
 
 import jakarta.validation.constraints.NotNull;
 import org.springframework.hateoas.EntityModel;
