+++ src/main/resources/application.yaml
@@ -1,3 +1,5 @@
+# Properties spring boot https://docs.spring.io/spring-boot/appendix/application-properties/index.html
+
 spring:
   application:
     name: medical-api
